requirements.txt
fastapi[standard]
masonite-orm # masonite-orm==2.18.6
psycopg2-binary # psycopg2-binary==2.9.5
strawberry-graphql[fastapi] # strawberry-graphql[fastapi]==0.158.0
